# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: |
    Omninity Operator HTTP API.

    Every 2xx response conforms to the canonical envelope shape defined in
    `Standard 1` of the bug-prevention standards: `{ success, data, error }`.
    Collection endpoints further wrap their payload in
    `{ items, nextCursor }` per `Standard 13`.

    Tenant context is supplied via the `X-Tenant-ID` header until session
    cookies replace it on `/api/auth/login`.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Bootstrap a new tenant + owner user
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: User created and signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Sign in with email + password
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in; session cookie set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Return the signed-in user (singleton)
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /agent/runs:
    get:
      operationId: listAgentRuns
      tags: [agent]
      summary: List recent agent runs in this workspace
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Agent runs page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunListResponse"
    post:
      operationId: createAgentRun
      tags: [agent]
      summary: Kick off a new deterministic agent run
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRunRequest"
      responses:
        "200":
          description: Agent run created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunResponse"
  /agent/runs/{id}:
    get:
      operationId: getAgentRun
      tags: [agent]
      summary: Fetch one agent run by id
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent run details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunResponse"
        "404":
          description: Unknown run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /agent/runs/{id}/cancel:
    post:
      operationId: cancelAgentRun
      tags: [agent]
      summary: Cancel a running agent
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunResponse"
  /agent/runs/{id}/messages:
    get:
      operationId: listAgentRunMessages
      tags: [agent]
      summary: Conversation messages for one run
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Messages page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"
  /agent/runs/{id}/tool-calls:
    get:
      operationId: listAgentRunToolCalls
      tags: [agent]
      summary: Tool calls executed inside one run
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Tool calls page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolCallListResponse"
  /agent/runs/{id}/approvals:
    get:
      operationId: listAgentRunApprovals
      tags: [agent]
      summary: Approval gates for one run
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Approvals page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalListResponse"
  /agent/approvals/{id}/decide:
    post:
      operationId: decideApproval
      tags: [agent]
      summary: Approve or deny a pending approval gate
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
      responses:
        "200":
          description: Decision recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalResponse"
...
  /browser/screenshot:
    post:
      operationId: browserScreenshot
      tags: [browser]
      summary: Capture a screenshot via the local browser tool (Tier 1 stub)
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrowserScreenshotRequest"
      responses:
        "200":
          description: Screenshot scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrowserActionResponse"
  /browser/extract:
    post:
      operationId: browserExtract
      tags: [browser]
      summary: Extract structured content from a page (Tier 1 stub)
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrowserExtractRequest"
      responses:
        "200":
          description: Extraction scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrowserActionResponse"
components:
  parameters:
    TenantHeader:
      name: X-Tenant-ID
      in: header
      required: true
      description: |
        Tenant identifier. Replaced by JWT-derived context once full SSO
        ships — until then this header is the request's tenant context.
      schema:
        type: string
    CursorParam:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned by the previous page.
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      required: false
      description: Page size, default 20, max 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ApiErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          $ref: "#/components/schemas/ApiError"
...
    RegisterRequest:
      type: object
      required: [email, password, displayName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
        displayName:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthUser:
      type: object
      required: [id, email, displayName, role]
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
    AuthSession:
      type: object
      required: [user, expiresAt]
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
        expiresAt:
          type: string
          format: date-time
    AuthSessionResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/AuthSession"
...
    AgentRun:
      type: object
      required: [id, goal, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        goal:
          type: string
        status:
          type: string
        plan:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        modelName:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AgentRunListPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AgentRun"
        nextCursor:
          type: string
          nullable: true
    AgentRunListResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/AgentRunListPage"
    AgentRunResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/AgentRun"
    CreateAgentRunRequest:
      type: object
      required: [goal]
      properties:
        goal:
          type: string
        modelName:
          type: string
    Message:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id:
          type: string
        runId:
          type: string
          nullable: true
        role:
          type: string
        content:
          type: string
        tokensIn:
          type: integer
          nullable: true
        tokensOut:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
    MessageListPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        nextCursor:
          type: string
          nullable: true
    MessageListResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/MessageListPage"
    ToolCall:
      type: object
      required: [id, runId, toolName, status, riskLevel, createdAt]
      properties:
        id:
          type: string
        runId:
          type: string
        toolName:
          type: string
        status:
          type: string
        riskLevel:
          type: string
        input:
          type: string
        output:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        durationMs:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
    ToolCallListPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ToolCall"
        nextCursor:
          type: string
          nullable: true
    ToolCallListResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/ToolCallListPage"
    Approval:
      type: object
      required: [id, runId, toolCallId, reason, summary, decision, createdAt]
      properties:
        id:
          type: string
        runId:
          type: string
        toolCallId:
          type: string
        reason:
          type: string
        summary:
          type: string
        decision:
          type: string
        decidedBy:
          type: string
          nullable: true
        decidedAt:
          type: string
          format: date-time
          nullable: true
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    ApprovalListPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Approval"
        nextCursor:
          type: string
          nullable: true
    ApprovalListResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/ApprovalListPage"
    ApprovalResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/Approval"
    ApprovalDecisionRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [approved, denied]
        note:
          type: string
...
    BrowserScreenshotRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        viewport:
          type: string
    BrowserExtractRequest:
      type: object
      required: [url, selector]
      properties:
        url:
          type: string
          format: uri
        selector:
          type: string
    BrowserActionReceipt:
      type: object
      required: [status, scheduledAt]
      properties:
        status:
          type: string
        scheduledAt:
          type: string
          format: date-time
        detail:
          type: string
    BrowserActionResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/BrowserActionReceipt"